# server/common/api.yml
swagger: "2.0"
info:
  version: 1.0.0
  title: kiln-api
  description: Node.js middleware to integrate Forms Foundry Kiln, ICM and communications layer.
basePath: /api
...
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /generateNewTemplate:
    post:
      tags:
        - Communications Layer
      summary: Generate a new form template in Comm Layer and save to ICM
      description: >
        Proxies the request to Communication Layer's POST /generateNewTemplate.
        The body should contain the same arguments accepted by the generate flow.
        On success, returns { errorCode: 0, message }. On failure, returns { error }.
      operationId: generateNewTemplate
      parameters:
        - in: body
          name: body
          description: JSON payload for generation
          required: true
          schema:
            type: object
            additionalProperties: true
            properties:
              attachmentId:
                type: string
                description: ICM attachment identifier (required)
                example: "1-4ZYB80E"
              formId:
                type: string
                description: Form template identifier (required)
                example: "CF8787"
              area:
                type: string
                description: Business area (validated against ICM metadata)
                example: "Service Request"
              CaseId:
                type: string
                description: Optional Case identifier
                example: "1-4ZYB34V"
              ContactId:
                type: string
                description: Optional Contact identifier
              SRId:
                type: string
                description: Optional Service Request identifier
            required:
              - attachmentId
              - formId
              - area
      responses:
        200:
          description: Form successfully generated
          schema:
            type: object
            required: [errorCode, message]
            properties:
              errorCode:
                type: integer
                example: 0
              message:
                type: string
                example: Successfully generated the form
        400:
          description: Invalid request or generation failure
          schema:
            type: object
            required: [error]
            properties:
              error:
                type: string
                example: The form cannot be generated.  